# Optimise query to populate notification statuses