@idavidmcdonald 2022-01-06
Any chance you could add pre and post EXPLAIN ANALYZEs for this please? Not essential but I think a good knowledge share