@idavidmcdonald 2022-01-06
Thanks for sharing the query plan! Useful for me!